refactor(sdk): extract enrollment + factory classmethods to _client/_enrollment.py
PR 9/N of the SDK split. Moves the entire enrollment + factory
classmethod surface (10 methods, ~870 LOC) out of
``cullis_sdk/client.py`` into a dedicated ``_EnrollmentMixin``
under ``cullis_sdk/_client/_enrollment.py``.

Factory classmethods:
* ``from_enrollment`` — bootstrap from Mastio enrollment URL.
* ``from_identity_dir`` — ADR-014 canonical runtime constructor
  (TLS client cert IS the credential).
* ``from_api_key_file`` — backwards-compat alias, DeprecationWarning
  preserved.
* ``from_connector`` — bootstrap from ~/.cullis/identity/ Desktop
  layout (cert+key+metadata.json+dpop.jwk+ca-chain.pem all preserved).
* ``from_user_principal_pem`` — ADR-021 PR4c shared-mode Frontdesk
  Ambassador factory (cert+key in per-process tempdir, principal_id
  4-segment parsing preserved byte-perfect).
* ``from_spiffe_workload_api`` — legacy SPIFFE→Court (ADR-011
  sunset, DeprecationWarning preserved).

Operator-side enrollment + helpers:
* ``enroll_via_byoca`` / ``enroll_via_spiffe`` — admin-side helpers.
* ``_do_enroll`` — shared machinery (DPoP keygen → POST → persist →
  build runtime client).
* ``_persist_enrollment`` (@staticmethod) — writes ``persist_to/
  {agent.json, dpop.jwk, cert.pem, key.pem}`` with 0600 on key.

Movement only — every ``cls.__new__(cls)`` + replica-attribute
pattern is preserved verbatim (memory feedback
``sdk_factory_init_skip``: factories bypass ``__init__`` and any
drift here reintroduces ``AttributeError``s downstream). No
``@deprecated`` decorator added — that's a separate PR after the
split is complete.

Implementation notes:
* The two instance methods ``proxy_headers`` + ``_egress_http``
  fall inside the same line range as the original ``from_*``
  block but are NOT factories — they stay in ``cullis_sdk.client``
  as before (used by Sessions / Messaging mixins).
* ``_check_insecure_tls`` and ``_build_proxy_http_client`` live in
  ``cullis_sdk.client`` (kept there in PR #722 for test backward
  compat). ``_enrollment.py`` imports them lazily inside each
  factory body — same pattern as ``cullis_sdk/_client/_rfq.py``
  — to break the import cycle with ``cullis_sdk.client``.
* Unused module-level imports purged from ``cullis_sdk/client.py``
  (``json``, ``sys``). ``json`` was only used by ``from_connector``
  /``_persist_enrollment``, ``sys`` was already orphan pre-refactor.

MRO order:
``class CullisClient(_AiGatewayMixin, _AuthMixin, _DiscoveryMixin,
_EnrollmentMixin, _GuardianMixin, _MessagingOneshotMixin,
_RfqMixin, _SessionsMixin, _WebSocketMixin)`` — ``E`` slots in
position 4, between ``_DiscoveryMixin`` and ``_GuardianMixin``.

Verification:
* ``pytest tests/ -n auto --dist=loadfile`` — **3501 passed**,
  38 skipped, 462s. The 2 failures are the pre-existing postgres
  binding flakes (``test_pg_session_and_signed_messages``,
  ``test_pg_nonce_replay_blocked``) documented in memory.
* ``./sandbox/smoke.sh full`` — **PASS=25 FAIL=0 SKIP=1**.
  Exercises the enrollment surface end-to-end:
  - B4.4 / B4.5 / B4.6 — agent-a / agent-b / byoca-a use
    ``from_identity_dir`` + DPoP login.
  - B4.7 — A2A messaging through enrolled identities.

Follows the pattern of #702 / #706 / #711 / #713 / #716 / #718 /

#720 / #722.
